pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-rally"
version = "0.1.0"
description = "MCP server for aggregating and analyzing CA Rally defects."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "Jon Kimmel"}
]
dependencies = [
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.25.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.3.0"
]

[project.scripts]
mcp-rally-server = "mcp_rally.server:main"

[tool.ruff]
line-length = 100
target-version = "py311"
